build: refactors and moving styles to their own file
