+++ spinning-svgs/index.html
@@ -4,34 +4,11 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Spinning SVGs</title>
-
-    <style>
-      body {
-        margin: 0;
-      }
-
-      .wrapper {
-        margin: 20px;
-        box-shadow: 0px 0px 8px 8px #ccc;
-        width: calc(99vw - 20px);
-        height: calc(98vh - 20px);
-      }
-
-      #canvas {
-        box-sizing: border-box;
-        width: 100%;
-        height: 100%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-      }
-    </style>
-
-    <script type="module" src="/src/index.ts"></script>
   </head>
   <body>
     <section class="wrapper">
       <div id="canvas">Thing</div>
     </section>
+    <script type="module" src="/src/index.ts"></script>
   </body>
 </html>
