+++ spinning-svgs/src/SpinningSVGs.css
@@ -1,0 +1,19 @@
+body {
+  margin: 0;
+}
+
+.wrapper {
+  margin: 20px;
+  box-shadow: 0px 0px 8px 8px #ccc;
+  width: calc(99vw - 20px);
+  height: calc(98vh - 20px);
+}
+
+#canvas {
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
