--- spinning-svgs/src/types.d.ts
@@ -1,7 +1,0 @@
-export interface GeneralSettings {
-  tag: string;
-  svg: string;
-  svgQuery: string;
-  svgWidth: number;
-  colours: Array<string>;
-}
